Refactors the interface and model loading to have more representative names
