+++ setup.cfg
@@ -21,4 +21,4 @@ include_package_data = True
 where=src
 
 [options.package_data]
-* = *.sht, *.yml, *.jar, *.pt, *.json, *.pdb
+* = *.sht, *.yml, *.jar, *.pt, *.json, *.pdb, *.npz
