+++ src/poli/core/util/proteins/rasp/__init__.py
@@ -1,2 +1,2 @@
-from .rasp import RaspInterface
-from .rasp import load_cavity_and_downstream_models
+from .rasp_interface import RaspInterface
+from .load_models import load_cavity_and_downstream_models
